Fixed not splitting in some levels in the PC port
Fixed the bug in which the auto splitter wasn't working properly in some levels of the PC port (Building Top Chase, Scale the Girders, Police Evaded, Spidey vs. Monster Ock).
I did a quick check and the auto-start feature still works for the PS1 version and it doesn't start when the Credits are playing.